% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence-check.R
\name{diagnostic_summary}
\alias{diagnostic_summary}
\title{Extract a Diagnostic Summary}
\usage{
diagnostic_summary(x)
}
\arguments{
\item{x}{An \code{hbmfit} or \code{hbcc_results} object.}
}
\value{
A named list of diagnostic statistics.
}
\description{
Extract a Diagnostic Summary
}
