% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-model.R
\name{get_hbsae_model}
\alias{get_hbsae_model}
\title{Inspect a Registered HBSAE Model Specification}
\usage{
get_hbsae_model(key)
}
\arguments{
\item{key}{Character.  A model key returned by
\code{\link{list_hbsae_models}}.}
}
\value{
The named list spec, or \code{NULL} if not found.
}
\description{
Inspect a Registered HBSAE Model Specification
}
\examples{
get_hbsae_model("lognormal")

}
\seealso{
\code{\link{register_hbsae_model}}
}
