% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbm_beta.R
\name{hbm_beta}
\alias{hbm_beta}
\title{Small Area Estimation using Hierarchical Bayesian under Beta Distribution}
\usage{
hbm_beta(response, 
                predictors, 
                n = NULL, 
                deff = NULL, 
                data, 
                prior = NULL)
}
\arguments{
\item{response}{The dependent (outcome) variable in the model. This variable represents the main response being predicted or analyzed.}

\item{predictors}{A list of independent (explanatory) variables used in the model. These variables form the fixed effects in the regression equation.}

\item{n}{The number of sample units for each region used in the survey}

\item{deff}{Design Effect}

\item{re}{Specifies the grouping variable for hierarchical (random effects) models. Used when modeling data with grouped structures (e.g., individuals within clusters).}

\item{sre}{An optional grouping factor mapping observations to spatial locations. If not specified, each observation is treated as a separate location. It is recommended to always specify a grouping factor to allow for handling of new data in postprocessing methods.}

\item{sre_type}{Determines the type of spatial random effect used in the model. The function currently supports "sar" and "car"}

\item{car_type}{Type of the CAR structure. Currently implemented are "escar" (exact sparse CAR), "esicar" (exact sparse intrinsic CAR), "icar" (intrinsic CAR), and "bym2".}

\item{sar_type}{Type of the SAR structure. Either "lag" (for SAR of the response values) or "error" (for SAR of the residuals).}

\item{M}{The M matrix in SAR is a spatial weighting matrix that shows the spatial relationship between locations with certain weights, while in CAR, the M matrix is an adjacency matrix that only contains 0 and 1 to show the proximity between locations. SAR is more focused on spatial influences with different intensities, while CAR is more on direct adjacency relationships. If sre is specified, the row names of M have to match the levels of the grouping factor}

\item{data}{Dataset used for model fitting}

\item{handle_missing}{Mechanism to handle missing data (NA values) to ensure model stability and prevent errors during estimation.}

\item{m}{Number of multiple imputations}

\item{prior}{Priors for the model parameters (default: \code{NULL})}

\item{control}{A list of control parameters for the sampler (default: \code{list()})}

\item{chains}{Number of Markov chains (default: 4)}

\item{iter}{Total number of iterations per chain (default: 4000)}

\item{warmup}{Number of warm-up iterations per chain (default: floor(iter/2))}

\item{cores}{Number of CPU cores to use (default: 1)}

\item{sample_prior}{(default: "no")}

\item{...}{}
}
\value{
A \code{hbmfit} object
}
\description{
This function is implemented a \strong{Hierarchical Bayesian Small Area Estimation (HBSAE)} model
under a \strong{beta distribution} using \strong{Bayesian inference} with the \code{brms} package.

The range of the variable data \eqn{(y)} that is intended as a beta distribution must be \eqn{0<y<1}.
The data proportion is supposed to be implemented with this function.

The function utilizes the \strong{Bayesian regression modeling framework} provided by \code{brms},
which interfaces with \strong{Stan} for efficient Markov Chain Monte Carlo (MCMC) sampling.
The \code{brm()} function from \code{brms} is used to estimate posterior distributions based on user-defined
hierarchical and spatial structures.
}
\examples{
\donttest{

# Prepare dataset
# Proses pengerjaan ya guys, belum menemukan data yang cocok bgt bisa n deff dan spatial 
}
}
