% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbm-info.R
\name{hbm_data}
\alias{hbm_data}
\title{Return the Data Used to Fit an hbmfit}
\usage{
hbm_data(model)
}
\arguments{
\item{model}{An \code{hbmfit} object.}
}
\value{
The original \code{data.frame} passed to the fitting function.
}
\description{
Return the Data Used to Fit an hbmfit
}
\examples{
\donttest{
library(hbsaems)
library(brms)
data("data_fhnorm")
model <- hbm(brms::bf(y ~ x1), data = data_fhnorm,
             re = ~ (1 | group),    # area-level random effect
             chains = 2, iter = 1000, warmup = 500,
             cores = 1, seed = 1, refresh = 0)
head(hbm_data(model))
}
}
