% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbm-info.R
\name{hbm_info}
\alias{hbm_info}
\title{Get Comprehensive Model Information}
\usage{
hbm_info(model)
}
\arguments{
\item{model}{An \code{hbmfit} object.}
}
\value{
A named list of metadata.
}
\description{
Returns a one-page summary of the fitted model's metadata: number of
observations, family, link function, formula, MCMC settings, missing-data
strategy, and so on.
}
\examples{
\donttest{
library(hbsaems)
library(brms)
data("data_fhnorm")
model <- hbm(brms::bf(y ~ x1), data = data_fhnorm,
             re = ~ (1 | regency),    # area-level random effect
             chains = 4, iter = 2000, warmup = 1000,
             cores = 1, seed = 1, refresh = 0)
hbm_info(model)
}
}
