% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbmc.R
\name{hbmc}
\alias{hbmc}
\title{hbmc: Check Model Goodness of Fit}
\usage{
hbmc(
  model,
  model2 = NULL,
  ndraws = 100,
  moment_match = FALSE,
  plot_types = c("pp_check", "params")
)
}
\arguments{
\item{model}{A \code{brmsfit} or \code{hbmfit}object.}

\item{model2}{An optional second \code{brmsfit} or \code{hbmfit} object object for model comparison.}

\item{ndraws}{Number of draws to plot in the posterior predictive check.}

\item{moment_match}{Logical; if \code{TRUE}, use moment matching for LOO computation.}

\item{plot_types}{Character vector specifying types of plots to generate (default: "pp_check", "params").}
}
\value{
A list containing model fit diagnostics:
\item{loo1}{LOO cross-validation diagnostics for \code{model}.}
\item{waic1}{WAIC value for \code{model}.}
\item{"pp_check"}{Posterior predictive check plot for assessing model fit for \code{model}.}
\item{"params"}{Plot of marginal posterior distributions for model parameters for \code{model}.}
\item{loo2}{(If \code{model2} is provided) LOO cross-validation diagnostics for \code{model2}.}
\item{waic2}{(If \code{model2} is provided) WAIC value for \code{model2}.}
\item{bf}{(If \code{model2} is provided) Bayes Factor comparing \code{model} and \code{model2} using bridge sampling.}
}
\description{
This function computes and prints the WAIC and LOO model fit diagnostics for a \code{brmsfit} object.
If a second model is provided, it also computes and prints the WAIC, LOO, and Bayes Factor for model comparison.
}
\details{
Check Model Goodness of Fit
}
\examples{
\donttest{

# Prepare dataset
data("BostonHousing")
data <- BostonHousing

# Fit the Basic Model
model <- hbm(
formula = bf(medv ~ crim + indus + rm + dis + rad + tax),  # Formula model
hb_sampling = "gaussian",      # Gaussian family for continuous outcomes
hb_link = "identity",          # Identity link function (no transformation)
data = data,                   # Dataset
chains = 4,                    # Number of MCMC chains
iter = 4000,                   # Total MCMC iterations
warmup = 2000,                 # Number of warmup iterations
cores = 2                      # Paralelisasi
)
summary(model)

# Model Goodness of Fit Checks
hbmc(model)

}
}
