% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbmfit-class.R
\name{is-hbsaems}
\alias{is-hbsaems}
\alias{is.hbmfit}
\alias{is.hbcc_results}
\alias{is.hbmc_results}
\alias{is.hbpc_results}
\alias{is.hbsae_results}
\title{Test Whether an Object Belongs to an hbsaems Result Class}
\usage{
is.hbmfit(x)

is.hbcc_results(x)

is.hbmc_results(x)

is.hbpc_results(x)

is.hbsae_results(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A single logical value.
}
\description{
Lightweight type-checking predicates for all five result classes produced
by \pkg{hbsaems}.  Each predicate returns a single logical.
}
\examples{
is.hbmfit("not a model")   # FALSE

}
