% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-model.R
\name{list_hbsae_models}
\alias{list_hbsae_models}
\title{List Registered HBSAE Models}
\usage{
list_hbsae_models()
}
\value{
Character vector of model keys.
}
\description{
Returns the keys of all model specs currently registered in the
\pkg{hbsaems} model registry.  Built-in models (\code{"gaussian"},
\code{"beta"}, \code{"binomial"}, \code{"lognormal"}, etc.) are always
included; user-registered models appear in addition.
}
\examples{
list_hbsae_models()

}
\seealso{
\code{\link{register_hbsae_model}}, \code{\link{hbm_flex}}
}
