% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Default Value for NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Any R objects.}
}
\value{
\code{x} if non-\code{NULL}, otherwise \code{y}.
}
\description{
Equivalent to the rlang \code{\%||\%} operator.  Internal use only.
}
\keyword{internal}
