% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbcc}
\alias{print.hbcc}
\title{Print a summary for a convergence check}
\usage{
\method{print}{hbcc}(x, ...)
}
\arguments{
\item{...}{Other potential arguments}

\item{object}{A \code{hbcc} object}
}
\value{
Summary of a convergence check
}
\description{
Print a summary for a convergence check
}
