% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbcc_results}
\alias{print.hbcc_results}
\title{Print a summary for a convergence check}
\usage{
\method{print}{hbcc_results}(x, ...)
}
\arguments{
\item{x}{A \code{hbcc} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a convergence check
}
\description{
Print a summary for a convergence check
}
