% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbmc}
\alias{print.hbmc}
\title{Print a summary for a model goodness of fit}
\usage{
\method{print}{hbmc}(x, ...)
}
\arguments{
\item{...}{Other potential arguments}

\item{object}{A \code{hbmc} object}
}
\value{
Summary of a model goodness of fit
}
\description{
Print a summary for a model goodness of fit
}
