% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbmc_results}
\alias{print.hbmc_results}
\title{Print a summary for a model goodness of fit and prior sensitivity}
\usage{
\method{print}{hbmc_results}(x, ...)
}
\arguments{
\item{x}{A \code{hbmc_results} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a model goodness of fit and prior sensitivity
}
\description{
Print a summary for a model goodness of fit and prior sensitivity
}
