% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbmfit}
\alias{print.hbmfit}
\title{Print a summary for a fitted model represented by a hbmfit object}
\usage{
\method{print}{hbmfit}(object, ...)
}
\arguments{
\item{object}{A \code{hbmfit} object}

\item{...}{Other potential arguments}
}
\value{
Summary of the model
}
\description{
Print a summary for a fitted model represented by a hbmfit object
}
