% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbpc_results}
\alias{print.hbpc_results}
\title{Print a summary for a prior predictive check}
\usage{
\method{print}{hbpc_results}(x, ...)
}
\arguments{
\item{x}{A \code{hbpc_results} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a prior predictive check
}
\description{
Print a summary for a prior predictive check
}
