% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbsae}
\alias{print.hbsae}
\title{Print a summary for a prediction result}
\usage{
\method{print}{hbsae}(x, ...)
}
\arguments{
\item{...}{Other potential arguments}

\item{object}{A \code{hbsae} object}
}
\value{
Summary of a prediction result
}
\description{
Print a summary for a prediction result
}
