% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.hbsae_results}
\alias{print.hbsae_results}
\title{Print a summary for a prediction result}
\usage{
\method{print}{hbsae_results}(x, ...)
}
\arguments{
\item{x}{A \code{hbsae} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a prediction result
}
\description{
Print a summary for a prediction result
}
