% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-methods.R
\name{prior_draws}
\alias{prior_draws}
\title{Extract Prior Draws}
\usage{
prior_draws(model, ...)
}
\arguments{
\item{model}{An \code{hbmfit} object.}

\item{...}{Additional arguments forwarded to
\code{\link[brms]{prior_draws}}.}
}
\value{
A \code{data.frame} of prior draws or \code{NULL} if no prior
  samples were stored in the model.
}
\description{
Requires the model to have been fit with \code{sample_prior = "yes"} or
\code{sample_prior = "only"}.
}
\examples{
\donttest{
library(hbsaems)
library(brms)
data("data_fhnorm")
model <- hbm(brms::bf(y ~ x1), data = data_fhnorm,
             re = ~ (1 | group),    # area-level random effect
             sample_prior = "yes",
             chains = 2, iter = 1000, warmup = 500,
             cores = 1, seed = 1, refresh = 0)
pd <- prior_draws(model)
head(pd)
}
}
