% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae-predict.R
\name{sae_filter}
\alias{sae_filter}
\title{Filter SAE Predictions by a Logical Condition}
\usage{
sae_filter(x, condition)
}
\arguments{
\item{x}{An \code{hbsae_results} object.}

\item{condition}{Logical vector of length equal to the number of areas.}
}
\value{
A new \code{hbsae_results} object containing only rows where
  \code{condition} is \code{TRUE}.
}
\description{
Filter SAE Predictions by a Logical Condition
}
\examples{
p <- structure(list(result_table = data.frame(Prediction = 1:5,
                                               RSE_percent = rep(5, 5)),
                     rse_model = 5, pred = 1:5),
                class = "hbsae_results")
sae_filter(p, p$pred > 2)

}
