% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae-predict.R
\name{sae_transform}
\alias{sae_transform}
\title{Apply a Transformation to SAE Predictions}
\usage{
sae_transform(x, fun, ...)
}
\arguments{
\item{x}{An \code{hbsae_results} object.}

\item{fun}{A function applied element-wise to the predictions.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A new \code{hbsae_results} object.
}
\description{
Apply a Transformation to SAE Predictions
}
\examples{
p <- structure(list(result_table = data.frame(Prediction = c(2, 4, 8),
                                               RSE_percent = c(5, 5, 5)),
                     rse_model = 5, pred = c(2, 4, 8)),
                class = "hbsae_results")
sae_transform(p, log)

}
