% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hbcc_results}
\alias{summary.hbcc_results}
\title{Create a summary of a convergence check}
\usage{
\method{summary}{hbcc_results}(object, ...)
}
\arguments{
\item{object}{A \code{hbcc} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a convergence check
}
\description{
Create a summary of a convergence check
}
