% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hbmc}
\alias{summary.hbmc}
\title{Create a summary of a model goodness of fit}
\usage{
\method{summary}{hbmc}(object, ...)
}
\arguments{
\item{object}{A \code{hbmc} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a model goodness of fit
}
\description{
Create a summary of a model goodness of fit
}
