% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hbmc_results}
\alias{summary.hbmc_results}
\title{Create a summary of a model goodness of fit and prior sensitivity}
\usage{
\method{summary}{hbmc_results}(object, ...)
}
\arguments{
\item{object}{A \code{hbmc_results} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a model goodness of fit and prior sensitivity
}
\description{
Create a summary of a model goodness of fit and prior sensitivity
}
