% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hbmfit}
\alias{summary.hbmfit}
\title{Create a summary of a fitted model represented by a hbmfit object}
\usage{
\method{summary}{hbmfit}(object, ...)
}
\arguments{
\item{object}{A \code{hbmfit} object}

\item{...}{Other potential arguments}
}
\value{
Summary of the model
}
\description{
Create a summary of a fitted model represented by a hbmfit object
}
