% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hbpc_results}
\alias{summary.hbpc_results}
\title{Create a summary of a prior predictive check}
\usage{
\method{summary}{hbpc_results}(object, ...)
}
\arguments{
\item{object}{A \code{hbpc_results} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a prior predictive check
}
\description{
Create a summary of a prior predictive check
}
