% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hbsae}
\alias{summary.hbsae}
\title{Create a summary of a prediction result}
\usage{
\method{summary}{hbsae}(object, ...)
}
\arguments{
\item{object}{A \code{hbsae} object}

\item{...}{Other potential arguments}
}
\value{
Summary of a prediction result
}
\description{
Create a summary of a prediction result
}
