% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-translations.R
\name{tr_langs}
\alias{tr_langs}
\title{List Available Languages}
\usage{
tr_langs()
}
\value{
Character vector of supported language codes.
}
\description{
List Available Languages
}
\examples{
tr_langs()

}
