% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjacency_matrix_car}
\alias{adjacency_matrix_car}
\title{Adjacency Matrix for Conditional Autoregressive Models}
\format{
A binary symmetric \eqn{5 \times 5} matrix with \code{1} for
  adjacent pairs of areas and \code{0} otherwise.
}
\source{
Simulated.
}
\usage{
adjacency_matrix_car
}
\description{
A small example adjacency matrix used for fitting Conditional
Autoregressive (CAR) random effects.
}
\keyword{datasets}
