% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbm.R
\name{hbm}
\alias{hbm}
\title{Hierarchical Bayesian Small Area Models}
\usage{
hbm(
  formula,
  sae_sampling = "gaussian",
  sae_link = "identity",
  re = NULL,
  sre = NULL,
  data,
  prior = NULL,
  control = list(),
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  cores = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{formula}{Formula specifying the model structure}

\item{sae_sampling}{The family distribution for the response}

\item{sae_link}{Link function for HBSAE}

\item{re}{Random effects formula}

\item{sre}{Spatial random effect formula with options for CAR or SAR}

\item{data}{Dataset used for model fitting}

\item{prior}{Priors for the model parameters (default: `NULL`)}

\item{control}{A list of control parameters for the sampler (default: `list()`)}

\item{chains}{Number of Markov chains (default: 4)}

\item{iter}{Total number of iterations per chain (default: 2000)}

\item{warmup}{Number of warm-up iterations per chain (default: floor(iter/2))}

\item{cores}{Number of CPU cores to use (default: 1)}

\item{seed}{Random seed for reproducibility (default: `NULL`)}

\item{...}{Additional arguments passed to `brm`}
}
\value{
A `brmsfit` object

A `brmsfit` object
#' @export
}
\description{
Hierarchical Bayesian Small Area Models
}
\details{
Hierarchical Bayesian Small Area Models
}
