% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbm_logitnormal.R
\name{hbm_logitnormal}
\alias{hbm_logitnormal}
\title{Small Area Estimation using Hierarchical Bayesian under Logit-Normal Distribution}
\usage{
hbm_logitnormal(
  response,
  trials,
  predictors,
  re = NULL,
  sre = NULL,
  sre_type = NULL,
  car_type = NULL,
  sar_type = NULL,
  M = NULL,
  data,
  handle_missing = NULL,
  m = 5,
  prior = NULL,
  control = list(),
  chains = 4,
  iter = 4000,
  warmup = floor(iter/2),
  cores = 1,
  sample_prior = "no",
  ...
)
}
\arguments{
\item{response}{The dependent (outcome) variable in the model. This variable represents the count of successes in a Binomial distribution.}

\item{trials}{Specifies the number of trials in a binomial model. This is required for binomial family models where the response variable is specified as a proportion.}

\item{predictors}{A list of independent (explanatory) variables used in the model. These variables form the fixed effects in the regression equation.}

\item{re}{Specifies the grouping variable for hierarchical (random effects) models. Used when modeling data with grouped structures (e.g., individuals within clusters).}

\item{sre}{An optional grouping factor mapping observations to spatial locations. If not specified, each observation is treated as a separate location. It is recommended to always specify a grouping factor to allow for handling of new data in postprocessing methods.}

\item{sre_type}{Determines the type of spatial random effect used in the model. The function currently supports "sar" and "car"}

\item{car_type}{Type of the CAR structure. Currently implemented are "escar" (exact sparse CAR), "esicar" (exact sparse intrinsic CAR), "icar" (intrinsic CAR), and "bym2".}

\item{sar_type}{Type of the SAR structure. Either "lag" (for SAR of the response values) or "error" (for SAR of the residuals).}

\item{M}{The M matrix in SAR is a spatial weighting matrix that shows the spatial relationship between locations with certain weights, while in CAR, the M matrix is an adjacency matrix that only contains 0 and 1 to show the proximity between locations. SAR is more focused on spatial influences with different intensities, while CAR is more on direct adjacency relationships. If sre is specified, the row names of M have to match the levels of the grouping factor}

\item{data}{Dataset used for model fitting}

\item{handle_missing}{Mechanism to handle missing data (NA values) to ensure model stability and prevent errors during estimation.}

\item{m}{Number of multiple imputations}

\item{prior}{Priors for the model parameters (default: \code{NULL})}

\item{control}{A list of control parameters for the sampler (default: \code{list()})}

\item{chains}{Number of Markov chains (default: 4)}

\item{iter}{Total number of iterations per chain (default: 2000)}

\item{warmup}{Number of warm-up iterations per chain (default: floor(iter/2))}

\item{cores}{Number of CPU cores to use (default: 1)}

\item{sample_prior}{(default: "no")}

\item{...}{}
}
\value{
A \code{hbmfit} object
}
\description{
This function implements a \strong{Hierarchical Bayesian Small Area Estimation (HBSAE)} model
under a \strong{logit-normal distribution} using \strong{Bayesian inference} with the \code{brms} package.

The model accounts for \strong{fixed effects}, \strong{random effects}, \strong{spatial random effects (CAR/SAR models)},
and \strong{measurement error correction}, allowing for robust small area estimation.

The function utilizes the \strong{Bayesian regression modeling framework} provided by \code{brms},
which interfaces with \strong{Stan} for efficient Markov Chain Monte Carlo (MCMC) sampling.
The \code{brm()} function from \code{brms} is used to estimate posterior distributions based on user-defined
hierarchical and spatial structures.
}
\examples{
\donttest{

# Prepare dataset
data("esoph")
esoph$n <- esoph$ncases + esoph$ncontrols


# Fit Logit-Normal Model
model1 <- hbm_logitnormal(
  response = "ncases",
  trials = "n",
  predictors = c("agegp", "tobgp"),
  data = esoph
)
summary(model1)


# Fit Logit-Normal Model With Missing Data
esoph[5:7, "ncases"] <- NA

model2 <- hbm_logitnormal(
  response = "ncases",
  trials = "n",
  predictors = c("agegp", "tobgp"),
  data = esoph,
  handle_missing = "deleted"
)
summary(model2)

model3 <- hbm_logitnormal(
  response = "ncases",
  trials = "n",
  predictors = c("agegp", "tobgp"),
  data = esoph,
  handle_missing = "multiple"
)
summary(model3)


# Fit Logit-Normal Model With Spatial Effect
# Create a Adjacency Matrix
esoph$area <- as.numeric(interaction(esoph$agegp, esoph$tobgp, esoph$alcgp, drop = TRUE))
n_areas <- length(unique(esoph$area))
adjacency_matrix <- bandSparse(n_areas, k = c(-1, 1), diag = list(rep(1, n_areas - 1), rep(1, n_areas - 1)))
area_names <- as.character(sort(unique(esoph$area)))
rownames(adjacency_matrix) <- area_names
colnames(adjacency_matrix) <- area_names

model4 <- hbm_logitnormal(
  response = "ncases",              # Response variable (number of successes)
  trials = "n",                     # Total trials (ncases + ncontrols)
  predictors = c("agegp", "tobgp", "alcgp"),  # Predictor variables
  sre = "area",                     # Spatial random effect based on the 'area' variable
  sre_type = "car",                 # Use Conditional Autoregressive (CAR) model for spatial effects
  M = adjacency_matrix,             # Adjacency matrix defining spatial relationships
  data = esoph                     # Input dataset
)
summary(model4)

} 

}
