% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_sae_app}
\alias{run_sae_app}
\title{Launch the Shiny App for Small Area Estimation using Hierarchical Bayesian}
\usage{
run_sae_app()
}
\value{
Opens a Shiny app in the web browser. The function does not return a value.
}
\description{
This function launches an interactive \strong{Shiny application} for performing
\strong{Hierarchical Bayesian Small Area Estimation (HBSAE)} using the \code{brms} package
with \code{Stan} as the backend for Bayesian inference.
}
