% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_weight_sar}
\alias{spatial_weight_sar}
\title{Spatial Weight Matrix for Simultaneous Autoregressive Models}
\format{
A \eqn{100 \times 100} numeric matrix with row sums equal to one
  (when at least one neighbour is present).
}
\source{
Simulated.
}
\usage{
spatial_weight_sar
}
\description{
A row-standardised spatial weight matrix for 100 areas, used for fitting
Simultaneous Autoregressive (SAR) random effects.
}
\keyword{datasets}
